fix: type error

+++ app/src/main/java/com/aicane/app/data/remote/dto/navigation/CreateInstructionRequest.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CreateInstructionRequest(
     val action: String,
-    val distanceMeters: Double,
+    val distanceMeters: Int,
     val message: String,
 )
